Move general util classes to their own types
