+++ extensions/barrage/src/main/java/io/deephaven/extensions/barrage/BarrageStreamGeneratorImpl.java
@@ -1169,88 +1169,6 @@ public class BarrageStreamGeneratorImpl implements BarrageStreamGenerator {
         }
     }
 
-    public static class DrainableByteArrayInputStream extends DefensiveDrainable {
-
-        private byte[] buf;
-        private final int offset;
-        private final int length;
-
-        public DrainableByteArrayInputStream(final byte[] buf, final int offset, final int length) {
-            this.buf = Objects.requireNonNull(buf);
-            this.offset = offset;
-            this.length = length;
-        }
-
-        @Override
-        public int available() {
-            if (buf == null) {
-                return 0;
-            }
-            return length;
-        }
-
-        @Override
-        public int drainTo(final OutputStream outputStream) throws IOException {
-            if (buf != null) {
-                try {
-                    outputStream.write(buf, offset, length);
-                } finally {
-                    buf = null;
-                }
-                return length;
-            }
-            return 0;
-        }
-    }
-
-    public static class ConsecutiveDrainableStreams extends DefensiveDrainable {
-        final DefensiveDrainable[] streams;
-
-        public ConsecutiveDrainableStreams(final @NotNull DefensiveDrainable... streams) {
-            this.streams = streams;
-        }
-
-        @Override
-        public int drainTo(final OutputStream outputStream) throws IOException {
-            int total = 0;
-            for (final DefensiveDrainable stream : streams) {
-                final int expected = total + stream.available();
-                total += ((Drainable) stream).drainTo(outputStream);
-                if (expected != total) {
-                    throw new IllegalStateException("drained message drained wrong number of bytes");
-                }
-                if (total < 0) {
-                    throw new IllegalStateException("drained message is too large; exceeds Integer.MAX_VALUE");
-                }
-            }
-            return total;
-        }
-
-        @Override
-        public int available() throws SizeException, IOException {
-            int total = 0;
-            for (final DefensiveDrainable stream : streams) {
-                total += stream.available();
-                if (total < 0) {
-                    throw new SizeException("drained message is too large; exceeds Integer.MAX_VALUE", total);
-                }
-            }
-            return total;
-        }
-
-        @Override
-        public void close() throws IOException {
-            for (final DefensiveDrainable stream : streams) {
-                try {
-                    stream.close();
-                } catch (final IOException e) {
-                    throw new UncheckedDeephavenException("unexpected IOException", e);
-                }
-            }
-            super.close();
-        }
-    }
-
     private static final class EmptyRowSetGenerator extends RowSetGenerator {
         public static final EmptyRowSetGenerator INSTANCE;
         static {
